updated cacheable portlet descriptor element for gs portlets
git-svn-id: http://svn.aei.mpg.de/eScience/GridSphere/trunk@2802 9c99c85f-4d0c-0410-8460-a9a1c48a3a7f

+++ src/org/gridlab/gridsphere/portletcontainer/impl/Cacheable.java
@@ -17,7 +17,7 @@ public class Cacheable {
         return shared;
     }
 
-    public void setExpiration(long seconds) {
+    public void setExpiration(int seconds) {
         this.seconds = seconds;
     }
